JavaScript Updated
Scroll Event:

Added a check to see if the menu is open (navbar.classList.contains('active')). If open, it closes the menu by removing the classes.
Menu Link Click:

Added an event listener to all links in the .navbar. When a link is clicked, the active and bx-x classes are removed, closing the menu.
Removed Scroll Toggling:

Removed unnecessary toggling of classes on scroll, which caused the random opening and vibration issue.